Centralize historical feed contracts
The feed-loader split left identical historical options, token normalization, and historical sort-key logic in multiple crates. Moving those contracts into ploy-market-contracts makes DB, Parquet, research, and collector paths share one type and one ordering rule.

Constraint: Review C1-C3 found incompatible duplicate types and duplicated business-critical EventDiscovered ordering

Rejected: Keep bridge conversion helpers between loader options | it keeps two incompatible public types alive

Confidence: high

Scope-risk: moderate

Directive: Put cross-crate feed DTOs and feed ordering helpers in ploy-market-contracts before adding new loaders

Tested: cargo test -p ploy-market-contracts --lib; cargo test -p ploy-feed-loaders --lib; cargo check -p ploy-feed-loaders --lib; cargo check -p ploy-strategy-bundles --no-default-features --lib; cargo check -p ploy-strategy-bundles --example run_backtest; cargo check -p ploy-research --no-default-features --features db --example factor_scan; cargo check -p ploy-market-data --lib

Not-tested: Parquet/DuckDB heavy feature execution with real parquet files
