+++ crates/ploy-feed-loaders/Cargo.toml
@@ -11,9 +11,9 @@ description = "Database-backed market-update loaders for Ploy"
 chrono.workspace = true
 ploy-market-contracts.workspace = true
 rust_decimal.workspace = true
-serde_json.workspace = true
 sqlx = { workspace = true, features = ["runtime-tokio-rustls", "postgres", "chrono", "rust_decimal", "json"] }
 tracing.workspace = true
 
 [dev-dependencies]
 rust_decimal_macros.workspace = true
+serde_json.workspace = true
