+++ crates/ploy-feed-loaders/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod database;
 
-pub use database::{load_from_database, load_from_database_with_options, HistoricalLoadOptions};
+pub use database::{load_from_database, load_from_database_with_options};
+pub use ploy_market_contracts::HistoricalLoadOptions;
