+++ crates/ploy-market-contracts/src/lib.rs
@@ -6,8 +6,8 @@ pub mod feed;
 pub mod instrument;
 pub mod venue;
 
-pub use events::MarketUpdate;
+pub use events::{market_update_sort_ts, normalize_token_id, MarketUpdate};
 pub use family::PredictionFamily;
-pub use feed::Feed;
+pub use feed::{Feed, HistoricalLoadOptions};
 pub use instrument::InstrumentKind;
 pub use venue::VenueKind;
