+++ crates/ploy-research/examples/factor_research.rs
@@ -239,29 +239,6 @@ async fn discover_valid_windows(
     .expect("valid window discovery failed")
 }
 
-fn market_update_ts(u: &MarketUpdate) -> DateTime<Utc> {
-    match u {
-        MarketUpdate::SpotPrice { ts, .. }
-        | MarketUpdate::AggTrade { ts, .. }
-        | MarketUpdate::Quote { ts, .. }
-        | MarketUpdate::L2 { ts, .. }
-        | MarketUpdate::L2Depth { ts, .. }
-        | MarketUpdate::SportsState { ts, .. }
-        | MarketUpdate::SportsPregame { ts, .. }
-        | MarketUpdate::SportsLive { ts, .. }
-        | MarketUpdate::ReferencePrice { ts, .. }
-        | MarketUpdate::Kline { ts, .. } => *ts,
-        MarketUpdate::EventDiscovered { end_time, window_secs, .. } => {
-            // Mirrors database.rs update_ts: subtract window + 1h buffer so EventDiscovered
-            // sorts before all quotes for the same event (quotes can arrive before start_time).
-            *end_time
-                - chrono::Duration::seconds(*window_secs as i64)
-                - chrono::Duration::hours(1)
-        }
-        MarketUpdate::EventExpired { end_time, .. } => *end_time,
-    }
-}
-
 /// Slices a sorted slice to items whose timestamp falls in `[start, end]`.
 /// Precondition: `items` must be sorted by `ts_fn` in ascending order.
 fn slice_by_time<'a, T>(
@@ -901,7 +878,7 @@ async fn main() {
     let mut total_event_rows = 0usize;
 
     for window in &windows {
-        // EventDiscovered sort-ts is end_time - window_secs - 1h (see market_update_ts).
+        // EventDiscovered sort-ts is end_time - window_secs - 1h.
         // Extend the lower bound by that same offset so EventDiscovered items are included.
         let updates_slice_start =
             window.start_time - chrono::Duration::hours(1) - chrono::Duration::seconds(300);
@@ -909,7 +886,7 @@ async fn main() {
             &all_updates,
             updates_slice_start,
             window.end_time,
-            market_update_ts,
+            MarketUpdate::sort_ts,
         );
 
         let lob_slice = slice_by_time(
