+++ crates/ploy-strategy-bundles/Cargo.toml
@@ -9,7 +9,7 @@ description = "Ploy strategy bundle runtime crate"
 
 [features]
 default = []
-parquet-feed = ["duckdb"]
+parquet-feed = ["dep:duckdb"]
 
 [dependencies]
 duckdb = { version = "1.1", features = ["chrono"], optional = true }
